refactor(team-mcp): extract shared TCP helpers to tcpHelpers.ts
Extract duplicated TCP message framing functions (writeTcpMessage,
createTcpMessageReader) and MCP script directory resolution logic
from TeamMcpServer and AionMcpService into a shared module.

- Create src/process/team/mcp/tcpHelpers.ts with 3 exported functions
- Update TeamMcpServer to import and use shared helpers
- Update AionMcpService to import and use shared helpers via @process alias
- Rename resolveTeamMcpDir/resolveAionMcpDir → resolveMcpScriptDir (unified)

Pure refactoring with zero behavior change. All tests pass.
