+++ src/process/team/TeamMcpServer.ts
@@ -19,6 +19,7 @@ import { isTeamCapableBackend, getTeamCapableBackends } from '@/common/types/tea
 import { ProcessConfig } from '@process/utils/initStorage';
 import { acpDetector } from '@process/agent/acp/AcpDetector';
 import { notifyMcpReady } from './mcpReadiness';
+import { writeTcpMessage, createTcpMessageReader, resolveMcpScriptDir } from './mcp/tcpHelpers';
 
 type SpawnAgentFn = (agentName: string, agentType?: string) => Promise<TeamAgent>;
 
@@ -39,63 +40,6 @@ export type StdioMcpConfig = {
   args: string[];
   env: Array<{ name: string; value: string }>;
 };
-
-// ── TCP message helpers ───────────────────────────────────────────────────────
-
-function writeTcpMessage(socket: net.Socket, data: unknown): void {
-  const json = JSON.stringify(data);
-  const body = Buffer.from(json, 'utf-8');
-  const header = Buffer.alloc(4);
-  header.writeUInt32BE(body.length, 0);
-  socket.write(Buffer.concat([header, body]));
-}
-
-function createTcpMessageReader(onMessage: (msg: unknown) => void): (chunk: Buffer) => void {
-  let buffer = Buffer.alloc(0);
-
-  return (chunk: Buffer) => {
-    buffer = Buffer.concat([buffer, chunk]);
-
-    while (buffer.length >= 4) {
-      const bodyLen = buffer.readUInt32BE(0);
-      if (buffer.length < 4 + bodyLen) break;
-
-      const jsonStr = buffer.subarray(4, 4 + bodyLen).toString('utf-8');
-      buffer = buffer.subarray(4 + bodyLen);
-
-      try {
-        const msg = JSON.parse(jsonStr);
-        onMessage(msg);
-      } catch {
-        // Malformed JSON — skip
-      }
-    }
-  };
-}
-
-/**
- * Resolve the directory containing the team-mcp-stdio.js bundle.
- * Mirrors the getBuiltinMcpBaseDir() logic in initStorage.ts so both MCP
- * scripts use the same path strategy across dev and packaged modes.
- *
- * In dev:       out/main/  (next to the main bundle)
- * In packaged:  app.asar.unpacked/out/main/  (asarUnpack makes it a real file)
- */
-function resolveTeamMcpDir(): string {
-  const mainModuleDir =
-    typeof require !== 'undefined' && require.main?.filename ? path.dirname(require.main.filename) : __dirname;
-  const baseDir = path.basename(mainModuleDir) === 'chunks' ? path.dirname(mainModuleDir) : mainModuleDir;
-  try {
-    // eslint-disable-next-line @typescript-eslint/no-require-imports
-    const { app } = require('electron');
-    if (app.isPackaged) {
-      return baseDir.replace('app.asar', 'app.asar.unpacked');
-    }
-  } catch {
-    // Not in Electron (unit tests / CLI mode) — use baseDir as-is
-  }
-  return baseDir;
-}
 
 /**
  * MCP server that provides team coordination tools to ACP agents.
@@ -148,7 +92,7 @@ export class TeamMcpServer {
    *   slot ID to every TCP request so the server knows who is calling.
    */
   getStdioConfig(agentSlotId?: string): StdioMcpConfig {
-    const scriptPath = path.join(resolveTeamMcpDir(), 'team-mcp-stdio.js');
+    const scriptPath = path.join(resolveMcpScriptDir(), 'team-mcp-stdio.js');
 
     const env: StdioMcpConfig['env'] = [
       { name: 'TEAM_MCP_PORT', value: String(this._port) },
